+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## Unreleased (v0.1.1)
+## Unreleased (v0.1.2)
 ### Added
 - Added accountType field to the AssetManager class.
 - Added a csv parser to convert csv strings into Assets, Books, Parties, Transactions. @JYL123
 - Added PrivateInvestment class (with ownership property).
 - Added ownership property to RealAsset and its subclasses.
 - Added `apiVersion` option to the api config. This will default to the latest stable version.
+- Added type declaration file for AssetManager class and api.AssetManagers.*
 
 ### Changed
 - Removed `maturityDate` from the Asset base class and set either maturityDate or expiryDate on the relevant subclasses.